Correct subtyping for records.
Use the '.' notation for records. This means that builtins are now stored in records, for instance

./liquidsoap -i 'l = list'
l     :
  [length:((['a])->int);mem:(('b,['b])->bool);rev:((['a])->['a]);
   remove:(('a,['a])->['a]);append:((['a],['a])->['a]);tl:((['a])->['a]);
   sort:(((('a,'a)->int),['a])->['a]);hd:(([string])->string);
   nth:((['a],int)->'a);fold:(((('a,'u)->'a),'a,['u])->'a);
   map:(((('a)->'u),['a])->['u]);iter:(((('a)->unit),['a])->unit)]
  where 'b is an orderable type

I did this with minimal changes to the code (as you can see, none of the plugins has to be adapted). There are still some TODO, which are mostly places where the code should be read again with special care (it should be ok but these things are tricky...).

--HG--
branch : LS-592