Make clock ticks abstract, implemented as bool
This is mostly meant to check an assumption I had about these ticks:
we only need to distinguish tic from tac! I have ran a simple example
involving smartcross, and it seems to work fine. More tests are in
order.

The only visible effect for users should be in get_clock_status: times
shown there are always going to be 0 or 1, much less useful information
to understand relative time flows.

If we ever merge this branch, we could keep the abstract tick type but
should move back to a less radical implementation.